refactor: Refactored `core.data` package (except AetherWorldProvider).
