+++ src/main/java/com/gildedgames/aether/Aether.java
@@ -160,7 +160,7 @@ public class Aether
 			generator.addProvider(new AetherItemTagData(generator, blockTags, helper));
 			generator.addProvider(new AetherEntityTagData(generator, helper));
 			generator.addProvider(new AetherFluidTagData(generator, helper));
-			generator.addProvider(new AetherAdvancementData(generator));
+			generator.addProvider(new AetherAdvancementData(generator, helper));
 			generator.addProvider(new AetherWorldData(generator));
 		}
 	}
