+++ src/main/java/com/gildedgames/aether/core/data/AetherItemTagData.java
@@ -187,7 +187,7 @@ public class AetherItemTagData extends ItemTagsProvider
                 .add(AetherItems.GOLDEN_RING.get())
                 .add(AetherItems.GOLDEN_PENDANT.get())
                 .add(AetherItems.GOLDEN_GLOVES.get());
-        tag(ItemTags.FOX_FOOD) //TODO: runData still needs to generate this.
+        tag(ItemTags.FOX_FOOD)
                 .add(AetherItems.BLUE_BERRY.get())
                 .add(AetherItems.ENCHANTED_BERRY.get());
         tag(ItemTags.MUSIC_DISCS)
