+++ src/main/java/com/gildedgames/aether/core/data/provider/SmartRegistryWriteOps.java
@@ -19,7 +19,8 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Function;
 
-public abstract class SmartRegistryWriteOps<Format> extends MultiRegistryWriteOps<Format> {
+public abstract class SmartRegistryWriteOps<Format> extends MultiRegistryWriteOps<Format>
+{
     protected HashCache directoryCache;
 
     public SmartRegistryWriteOps(String modId, DataGenerator generator, DynamicOps<Format> ops, Function<Format, String> fileWriter, RegistryAccess dynamicRegistries) {
@@ -30,7 +31,6 @@ public abstract class SmartRegistryWriteOps<Format> extends MultiRegistryWriteOp
     public String getName() {
         return "Aether World Data";
     }
-
 
     @Override
     public final void run(HashCache cache) {
